Abstracted communication adapter from state actor

+++ Kommander/Data/VoteRequest.cs
@@ -5,11 +5,11 @@ public sealed class VoteRequest
 {
     public int Partition { get; set; }
 
-    public int Term { get; set; }
+    public long Term { get; set; }
 
     public string Endpoint { get; set; }
 
-    public VoteRequest(int partition, int term, string endpoint)
+    public VoteRequest(int partition, long term, string endpoint)
     {
         Partition = partition;
         Term = term;
